+++ python/src/ilib/_ffi.py
@@ -53,6 +53,7 @@ IError ICopyImageScaled(IImage source, IImage dest, IGC gc, int src_x,
                         unsigned int src_height, int dest_x, int dest_y,
                         unsigned int dest_width, unsigned int dest_height);
 IError IReduceColors(IImage image, unsigned int max_colors);
+IError IDither(IImage image, unsigned int max_colors);
 
 IError IGreyscale(IImage image);
 IError INegate(IImage image);